fix(updateOptions): huge performance increase
